% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{CityWeather}
\alias{CityWeather}
\title{City_Weather_Class
Stores weather information for a given city.}
\usage{
CityWeather(city, temperature, humidity, pressure)
}
\description{
City_Weather_Class
Stores weather information for a given city.
}
