% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_lambda.R
\name{expected_lambda}
\alias{expected_lambda}
\title{Expected_attendance}
\usage{
expected_lambda(city_weather)
}
\arguments{
\item{city_weather}{A CityWeather object}
}
\value{
Expected Poisson rate λ
}
\description{
Expected_attendance
}
