% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{r_scalar}
\alias{r_scalar}
\title{Scalar radius evaluation for the glass profile}
\usage{
r_scalar(t, x1, x2, x3, x4, r_foot, r_stem, r_bowl, r_rim)
}
\arguments{
\item{t}{Height at which the radius is evaluated}

\item{x1, x2, x3, x4}{Region boundaries}

\item{r_foot, r_stem, r_bowl, r_rim}{Radii in different parts}
}
\value{
Radius at height \code{t}
}
\description{
Scalar radius evaluation for the glass profile
}
