% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{r_vec_for_integrate}
\alias{r_vec_for_integrate}
\title{Vectorized radius function}
\usage{
r_vec_for_integrate(t, x1, x2, x3, x4, r_foot, r_stem, r_bowl, r_rim)
}
\description{
Vectorized radius function
}
