% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius_fun.R
\name{S}
\alias{S}
\title{Easing_function_S(z)
Smooth transition function used to model glass curvature.}
\usage{
S(z)
}
\arguments{
\item{z}{Scaled value between 0 and 1}
}
\value{
Numeric value of easing function
}
\description{
Easing_function_S(z)
Smooth transition function used to model glass curvature.
}
