% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_party}
\alias{simulate_party}
\title{Simulate champagne consumption for a party}
\usage{
simulate_party(city_name, data, N = 1000, seed = 123L, bottle_L = 0.75)
}
\arguments{
\item{city_name}{Name of city}

\item{data}{Weather data list}

\item{N}{Number of simulation runs}

\item{seed}{Random seed}

\item{bottle_L}{Bottle capacity in liters}
}
\value{
List containing simulation outputs
}
\description{
Simulate champagne consumption for a party
}
\examples{
\dontrun{
data(list("processed_data"))
results <- simulate_party("Bern", processed_data)
}
}
