% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Glass.R
\name{volume_between.Glass_Profile}
\alias{volume_between.Glass_Profile}
\title{Volume between heights a and b for a glass}
\usage{
\method{volume_between}{Glass_Profile}(x)
}
\description{
Volume between heights a and b for a glass
}
